# docs/public/api/swagger-2.0.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Track, Manage, and Organize your Things.",
        "title": "Homebox API",
        "contact": {
            "name": "Homebox Team",
            "url": "https://discord.homebox.software"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/v1/group/backup-destinations": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Returns the group's backup destinations with schedule, retention and health status. Group owners only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backups"
                ],
                "summary": "List Backup Destinations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Results-repo_BackupDestinationOut"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Credentials are never stored: cloud destinations read them from the server environment.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backups"
                ],
                "summary": "Create a Backup Destination",
                "parameters": [
                    {
                        "description": "Destination settings",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repo.BackupInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/repo.BackupDestinationOut"
                        }
                    }
                }
            }
        },
        "/v1/group/backup-destinations/test": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backups"
                ],
                "summary": "Test Unsaved Backup Destination Settings",
                "parameters": [
                    {
                        "description": "Destination settings",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repo.BackupInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.TestResult"
                        }
                    }
                }
            }
        },
        "/v1/group/backup-destinations/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backups"
                ],
                "summary": "Get a Backup Destination",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Destination ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repo.BackupDestinationOut"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backups"
                ],
                "summary": "Update a Backup Destination",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Destination ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Destination settings",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repo.BackupInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repo.BackupDestinationOut"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Removes the destination and its backup history. Files already written to the destination are not deleted.",
                "tags": [
                    "Backups"
                ],
                "summary": "Delete a Backup Destination",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Destination ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/v1/group/backup-destinations/{id}/run": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Starts an on-demand backup to this destination.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backups"
                ],
                "summary": "Back Up Now",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Destination ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/repo.ExportOut"
                        }
                    }
                }
            }
        },
        "/v1/group/backup-destinations/{id}/test": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Writes and deletes a small test object and records the result as the destination's health.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backups"
                ],
                "summary": "Test a Saved Backup Destination",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Destination ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.TestResult"
                        }
                    }
                }
            }
        },
        "/v1/group/backup-destinations/{id}/versions": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Returns the backups written to a destination, newest first.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backups"
                ],
                "summary": "List Backup Versions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Destination ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Results-repo_ExportOut"
                        }
                    }
                }
            }
        },
        "/v1/group/backup-options": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backups"
                ],
                "summary": "Get Backup Options",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.BackupOptions"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "repo.BackupDestinationOut": {
            "type": "object",
            "required": [
                "frequency",
                "name",
                "type"
            ],
            "properties": {
                "alertFailureThreshold": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 1
                },
                "alertStaleHours": {
                    "description": "AlertStaleHours alerts when no backup succeeded for this long. 0 disables.",
                    "type": "integer",
                    "maximum": 8760,
                    "minimum": 0
                },
                "alertsEnabled": {
                    "type": "boolean"
                },
                "atHour": {
                    "type": "integer",
                    "maximum": 23,
                    "minimum": 0
                },
                "atMinute": {
                    "type": "integer",
                    "maximum": 59,
                    "minimum": 0
                },
                "connString": {
                    "type": "string",
                    "maxLength": 2048
                },
                "createdAt": {
                    "type": "string"
                },
                "dayOfMonth": {
                    "type": "integer",
                    "maximum": 28,
                    "minimum": 1
                },
                "description": {
                    "type": "string",
                    "maxLength": 1000
                },
                "enabled": {
                    "type": "boolean"
                },
                "frequency": {
                    "description": "Frequency is one of hourly, daily, weekly, monthly.",
                    "type": "string",
                    "enum": [
                        "hourly",
                        "daily",
                        "weekly",
                        "monthly"
                    ]
                },
                "groupId": {
                    "type": "string"
                },
                "hasSecret": {
                    "description": "HasSecret reports whether credentials are stored for the destination.",
                    "type": "boolean"
                },
                "healthCheckedAt": {
                    "type": "string",
                    "x-nullable": true
                },
                "healthError": {
                    "type": "string"
                },
                "healthFailures": {
                    "type": "integer"
                },
                "healthIntervalMinutes": {
                    "type": "integer",
                    "maximum": 1440,
                    "minimum": 1
                },
                "healthStatus": {
                    "type": "string"
                },
                "hostKey": {
                    "type": "string",
                    "maxLength": 255
                },
                "id": {
                    "type": "string"
                },
                "intervalHours": {
                    "type": "integer",
                    "maximum": 168,
                    "minimum": 1
                },
                "keepDaily": {
                    "type": "integer",
                    "maximum": 3650,
                    "minimum": 0
                },
                "keepMonthly": {
                    "type": "integer",
                    "maximum": 120,
                    "minimum": 0
                },
                "keepWeekly": {
                    "type": "integer",
                    "maximum": 520,
                    "minimum": 0
                },
                "lastError": {
                    "type": "string"
                },
                "lastRunAt": {
                    "type": "string",
                    "x-nullable": true
                },
                "lastSkippedAt": {
                    "type": "string",
                    "x-nullable": true
                },
                "lastSuccessAt": {
                    "type": "string",
                    "x-nullable": true
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "nextRunAt": {
                    "type": "string",
                    "x-nullable": true
                },
                "prefix": {
                    "type": "string",
                    "maxLength": 255
                },
                "scheduleEnabled": {
                    "type": "boolean"
                },
                "skipIfUnchanged": {
                    "type": "boolean"
                },
                "type": {
                    "description": "Type is one of primary, local, s3, gcs, azblob, sftp, webdav.",
                    "type": "string",
                    "enum": [
                        "primary",
                        "local",
                        "s3",
                        "gcs",
                        "azblob",
                        "sftp",
                        "webdav"
                    ]
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "description": "Username and HostKey serve the sftp and webdav types. HostKey is the\nSSH host key fingerprint (SHA256:...) an sftp server must present.",
                    "type": "string",
                    "maxLength": 255
                },
                "weekday": {
                    "type": "integer",
                    "maximum": 6,
                    "minimum": 0
                }
            }
        },
        "repo.BackupInput": {
            "type": "object",
            "required": [
                "frequency",
                "name",
                "type"
            ],
            "properties": {
                "alertFailureThreshold": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 1
                },
                "alertStaleHours": {
                    "description": "AlertStaleHours alerts when no backup succeeded for this long. 0 disables.",
                    "type": "integer",
                    "maximum": 8760,
                    "minimum": 0
                },
                "alertsEnabled": {
                    "type": "boolean"
                },
                "atHour": {
                    "type": "integer",
                    "maximum": 23,
                    "minimum": 0
                },
                "atMinute": {
                    "type": "integer",
                    "maximum": 59,
                    "minimum": 0
                },
                "connString": {
                    "type": "string",
                    "maxLength": 2048
                },
                "dayOfMonth": {
                    "type": "integer",
                    "maximum": 28,
                    "minimum": 1
                },
                "description": {
                    "type": "string",
                    "maxLength": 1000
                },
                "destinationId": {
                    "description": "DestinationID is only used when testing unsaved settings: it lets the\ntest reuse the stored credentials of that destination.",
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "frequency": {
                    "description": "Frequency is one of hourly, daily, weekly, monthly.",
                    "type": "string",
                    "enum": [
                        "hourly",
                        "daily",
                        "weekly",
                        "monthly"
                    ]
                },
                "healthIntervalMinutes": {
                    "type": "integer",
                    "maximum": 1440,
                    "minimum": 1
                },
                "hostKey": {
                    "type": "string",
                    "maxLength": 255
                },
                "intervalHours": {
                    "type": "integer",
                    "maximum": 168,
                    "minimum": 1
                },
                "keepDaily": {
                    "type": "integer",
                    "maximum": 3650,
                    "minimum": 0
                },
                "keepMonthly": {
                    "type": "integer",
                    "maximum": 120,
                    "minimum": 0
                },
                "keepWeekly": {
                    "type": "integer",
                    "maximum": 520,
                    "minimum": 0
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "password": {
                    "description": "Password is the sftp or webdav password. Leave empty to keep the stored\ncredentials of an existing destination.",
                    "type": "string"
                },
                "prefix": {
                    "type": "string",
                    "maxLength": 255
                },
                "privateKey": {
                    "description": "PrivateKey is an unencrypted PEM private key for sftp.",
                    "type": "string"
                },
                "scheduleEnabled": {
                    "type": "boolean"
                },
                "skipIfUnchanged": {
                    "type": "boolean"
                },
                "type": {
                    "description": "Type is one of primary, local, s3, gcs, azblob, sftp, webdav.",
                    "type": "string",
                    "enum": [
                        "primary",
                        "local",
                        "s3",
                        "gcs",
                        "azblob",
                        "sftp",
                        "webdav"
                    ]
                },
                "username": {
                    "description": "Username and HostKey serve the sftp and webdav types. HostKey is the\nSSH host key fingerprint (SHA256:...) an sftp server must present.",
                    "type": "string",
                    "maxLength": 255
                },
                "weekday": {
                    "type": "integer",
                    "maximum": 6,
                    "minimum": 0
                }
            }
        },
...
        "repo.ExportOut": {
            "type": "object",
            "properties": {
                "artifactPath": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "destinationId": {
                    "description": "DestinationID is the backup destination holding the artifact; absent\nfor plain manual exports, which live in the primary storage.",
                    "type": "string",
                    "x-nullable": true
                },
                "error": {
                    "type": "string"
                },
                "groupId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "kind": {
                    "description": "Kind is \"export\" for server-produced backup artifacts, \"import\" for\nuser-uploaded restore zips. The lifecycle fields below behave the\nsame for both.",
                    "type": "string"
                },
                "origin": {
                    "description": "Origin is \"manual\" for user-triggered exports and \"scheduled\" for\nbackups started by the scheduler.",
                    "type": "string"
                },
                "progress": {
                    "type": "integer"
                },
                "sizeBytes": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "services.TestResult": {
            "type": "object",
            "properties": {
                "hostKey": {
                    "description": "HostKey is the SSH host key fingerprint the server presented when it\nwas missing or did not match, so the UI can offer to trust it.",
                    "type": "string"
                },
                "latencyMs": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        },
...
        "v1.BackupOptions": {
            "type": "object",
            "properties": {
                "allowCustomEndpoints": {
                    "type": "boolean"
                },
                "enabled": {
                    "type": "boolean"
                },
                "localEnabled": {
                    "type": "boolean"
                },
                "remoteEnabled": {
                    "type": "boolean"
                }
            }
        },
...
        "v1.Results-repo_BackupDestinationOut": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repo.BackupDestinationOut"
                    }
                }
            }
        },
        "v1.Results-repo_ExportOut": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/repo.ExportOut"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "\"Type 'Bearer TOKEN' to correctly set the API Key\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}